requirements.txt
aniso8601==1.3.0
asn1crypto==0.22.0
Beaker==1.5.4
beautifulsoup4==4.6.0
blivet==2.1.11
blivet-gui==2.1.7
Brlapi==0.6.6
cffi==1.9.1
chardet==2.3.0
click==6.7
coverage==4.4.1
cryptography==2.0.2
cssselect==0.9.2
cupshelpers==1.0
dateutils==0.6.6
decorator==4.0.11
Flask==0.12.2
Flask-RESTful==0.3.6
fros==1.1
gpg==1.8.0
html5lib==0.999
humanize==0.5.1
idna==2.5
iniparse==0.4
initial-setup==0.3.46
IPy==0.81
ipython==6.2.1
ipython-genutils==0.2.0
isc==2.0
itsdangerous==0.24
javapackages==4.7.0
jedi==0.11.0
Jinja2==2.9.6
langtable==0.0.37
lxml==3.7.2
Mako==1.0.6.dev0
MarkupSafe==0.23
marshmallow==2.13.6
ntplib==0.3.3
olefile==0.44
ordered-set==2.0.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
pid==2.1.1
Pillow==4.1.1
ply==3.9
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pwquality==1.3.0
pyasn1==0.2.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0
pyenchant==1.6.10
Pygments==2.2.0
pygobject==3.24.1
pyinotify==0.9.6
PyIscsi==1.0
pykickstart==2.35
pyOpenSSL==16.2.0
pyparsing==2.1.10
pyparted==3.10.7
PySocks==1.6.7
python-augeas==0.5.0
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-meh==0.43
pytz==2016.10
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.13.0
requests-file==1.4
requests-ftp==0.3.1
rpm-python==4.13.0.1
sepolicy==1.1
setools==4.1.0
setroubleshoot==1.1
simplegeneric==0.8.1
six==1.10.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.4
SQLAlchemy==1.1.14
SQLAlchemy-Utils==0.32.18
SSSDConfig==1.15.3
systemd-python==234
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
webargs==1.8.1
Werkzeug==0.12.2
wrapt==1.10.10
